requirements.txt
keras
pillow
scipy
tensorflow